BhyvePkg: fix BhyveSetGraphicsMode call for VS2019 build
The VS2019 toolchain warns about casts from UINT32 to UINT16.
Work around this by adding casts in the call to BhyveSetGraphicsMode, as
a temporary measure.

Message-Id: <9289b41b8656641febae54b560c7aafe59ef2aca.1586991816.git.rebecca@bsdio.com>

+++ BhyvePkg/BhyveRfbDxe/GopScreen.c
@@ -184,7 +184,7 @@ EmuGopSetMode (
   This->Mode->FrameBufferSize = Private->FbSize;
   DEBUG ((DEBUG_INFO, "BHYVE GOP FrameBufferBase: 0x%x, FrameBufferSize: 0x%x\n", This->Mode->FrameBufferBase, This->Mode->FrameBufferSize));
 
-  BhyveSetGraphicsMode(Private, ModeData->HorizontalResolution, ModeData->VerticalResolution, ModeData->ColorDepth);
+  BhyveSetGraphicsMode(Private, (UINT16)ModeData->HorizontalResolution, (UINT16)ModeData->VerticalResolution, (UINT16)ModeData->ColorDepth);
 
   RETURN_STATUS ret = FrameBufferBltConfigure (
     (VOID*)(UINTN) This->Mode->FrameBufferBase,
